tidy up unused imports and format once again
