+++ packages/react-client/src/types.ts
@@ -59,10 +59,6 @@ export type Errors = {
   video?: DeviceError | null;
 };
 
-export type GetMedia =
-  | { stream: MediaStream; type: "OK"; constraints: MediaStreamConstraints; previousErrors: Errors }
-  | { error: DeviceError | null; type: "Error"; constraints: MediaStreamConstraints };
-
 export type CurrentDevices = { videoinput: MediaDeviceInfo | null; audioinput: MediaDeviceInfo | null };
 
 export type UseSetupMediaConfig<TrackMetadata> = {
@@ -198,30 +194,6 @@ export type Devices<TrackMetadata> = {
   camera: UserMediaAPI<TrackMetadata>;
   microphone: UserMediaAPI<TrackMetadata>;
   screenShare: ScreenShareAPI<TrackMetadata>;
-};
-
-export const PERMISSION_DENIED: DeviceError = { name: "NotAllowedError" };
-export const OVERCONSTRAINED_ERROR: DeviceError = { name: "OverconstrainedError" };
-export const NOT_FOUND_ERROR: DeviceError = { name: "NotFoundError" };
-export const UNHANDLED_ERROR: DeviceError = { name: "UNHANDLED_ERROR" };
-
-// https://developer.mozilla.org/en-US/docs/Web/API/MediaDevices/getUserMedia#exceptions
-// OverconstrainedError has higher priority than NotAllowedError
-export const parseUserMediaError = (error: unknown): DeviceError | null => {
-  if (!(error instanceof DOMException)) {
-    console.warn({ name: "Unhandled getUserMedia error", error });
-    return null;
-  }
-
-  if (error.name === "NotAllowedError") {
-    return PERMISSION_DENIED;
-  } else if (error.name === "OverconstrainedError") {
-    return OVERCONSTRAINED_ERROR;
-  } else if (error.name === "NotFoundError") {
-    return NOT_FOUND_ERROR;
-  }
-
-  return null;
 };
 
 export type FishjamContextProviderProps = {
